refactor: give the Node services one OpenTelemetry setup
The relay's copy was made deliberately — six lines of Telegram guard did
not justify a package — on the stated premise that the relay makes no
outbound HTTP. That premise was wrong: it reaches Slack through the SDK,
not a bare fetch, and once traces were flowing on test it showed 80 calls
to slack.com/api/auth.test at p95 697ms, every one of them named bare
POST in the waterfall.

So the relay needs undiciClientSpanName too, and copying it would put
~90 identical lines in two places — including the URL-redaction
heuristics, which are the ones that must not drift. A fix to redaction in
one copy would silently not reach the other, and the failure mode is a
credential in an exported, retained span name.

@agentconnect.md/observability now owns the SDK wiring, the
enable/disable predicate, and both hygiene functions. Each service keeps
only what genuinely differs: its name, its version (resolved by the
caller, since the shared package would otherwise report its own), and any
instrumentation only it has — Prisma, for the control plane. The two
service modules go from 197 and 142 lines to 56 and 42, and re-export the
same names, so no call site changes.

The relay gains client span naming as a side effect of the move, which is
the reason for it.

Also verified beyond the test suites, since a new workspace package
touches things tests do not reach: both container images build, and
@agentconnect.md/observability resolves and loads inside the relay image.
The Dockerfile enumerates packages explicitly, so the manifests stage and
both topological build lists needed the new one.
