+++ CLAUDE.md
@@ -42,18 +42,19 @@ callback endpoint is required; the CP only orchestrates. Concretely:
 
 ## Monorepo (pnpm workspace, `packages/*`)
 
-| Package                               | Stack                              | Role                                                                                                                                                               |
-| ------------------------------------- | ---------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| `@agentconnect.md/message`            | TypeScript                         | Pure Slack, Lark, Telegram, and Discord normalization; depends on protocol types and contains no platform SDKs or I/O.                                             |
-| `@agentconnect.md/protocol`           | zod                                | Shared wire contract — frames, normalized message schemas, and fencing fields (`sessionEpoch`/`seq`/`launchId`). Single source of truth for every wire consumer.   |
-| `@agentconnect.md/connection`         | TypeScript                         | Wire-agnostic WS dial-out/accept primitives (transport, correlator, backoff, keepalive) shared by daemon, relay, and control-plane.                                |
-| `@agentconnect.md/cli`                | Node CLI (commander)               | The stable `agentconnect` bin: daemon lifecycle (`up`/`down`/`restart`/`status`), service install, `login`, and version management/upgrades.                       |
-| `@agentconnect.md/daemon`             | Node CLI (commander)               | The edge unit. Exposes the `agentconnect-daemon` bin (`run`, `chat`, `agent list`, plus hidden `mcp-bridge`/`git-credential`/`gh-token` helpers).                  |
-| `@agentconnect.md/control-plane`      | Fastify + Prisma (Postgres)        | One Fastify process co-hosts the C2 BFF REST surface **and** the daemon WS endpoint on one port / one Postgres connection.                                         |
-| `@agentconnect.md/relay`              | Fastify                            | Optional public ingress: Slack/Feishu HTTP callbacks, webhooks, webchat. Verifies, demuxes to the owning bot, forwards to the daemon. Persists no message content. |
-| `@agentconnect.md/setup`              | Fastify + browser UI               | Loopback-only Setup Server for self-hosting, Logto setup, and deployment-owned provider App administration.                                                        |
-| `@agentconnect.md/web`                | Next.js 16 + React 19 + Tailwind 4 | Config / monitoring console.                                                                                                                                       |
-| `@agentconnect.md/memory-plugin-mem0` | TypeScript                         | Memory-plugin profile wrappers for Mem0 Cloud and OSS.                                                                                                             |
+| Package                               | Stack                              | Role                                                                                                                                                                                                         |
+| ------------------------------------- | ---------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| `@agentconnect.md/message`            | TypeScript                         | Pure Slack, Lark, Telegram, and Discord normalization; depends on protocol types and contains no platform SDKs or I/O.                                                                                       |
+| `@agentconnect.md/protocol`           | zod                                | Shared wire contract — frames, normalized message schemas, and fencing fields (`sessionEpoch`/`seq`/`launchId`). Single source of truth for every wire consumer.                                             |
+| `@agentconnect.md/connection`         | TypeScript                         | Wire-agnostic WS dial-out/accept primitives (transport, correlator, backoff, keepalive) shared by daemon, relay, and control-plane.                                                                          |
+| `@agentconnect.md/observability`      | OpenTelemetry                      | The Node services' shared OTel bootstrap plus the span-name hygiene that keeps credentials out of exported names. Control-plane and relay each pass only their own name, version, and extra instrumentation. |
+| `@agentconnect.md/cli`                | Node CLI (commander)               | The stable `agentconnect` bin: daemon lifecycle (`up`/`down`/`restart`/`status`), service install, `login`, and version management/upgrades.                                                                 |
+| `@agentconnect.md/daemon`             | Node CLI (commander)               | The edge unit. Exposes the `agentconnect-daemon` bin (`run`, `chat`, `agent list`, plus hidden `mcp-bridge`/`git-credential`/`gh-token` helpers).                                                            |
+| `@agentconnect.md/control-plane`      | Fastify + Prisma (Postgres)        | One Fastify process co-hosts the C2 BFF REST surface **and** the daemon WS endpoint on one port / one Postgres connection.                                                                                   |
+| `@agentconnect.md/relay`              | Fastify                            | Optional public ingress: Slack/Feishu HTTP callbacks, webhooks, webchat. Verifies, demuxes to the owning bot, forwards to the daemon. Persists no message content.                                           |
+| `@agentconnect.md/setup`              | Fastify + browser UI               | Loopback-only Setup Server for self-hosting, Logto setup, and deployment-owned provider App administration.                                                                                                  |
+| `@agentconnect.md/web`                | Next.js 16 + React 19 + Tailwind 4 | Config / monitoring console.                                                                                                                                                                                 |
+| `@agentconnect.md/memory-plugin-mem0` | TypeScript                         | Memory-plugin profile wrappers for Mem0 Cloud and OSS.                                                                                                                                                       |
 
 When you change a frame in `protocol`, both daemon and CP consume it — rebuild
 `protocol` (or rely on its `development` export → `./src/index.ts`) and check both sides.
